citizen snippets
