+++ wpilibjExamples/src/main/java/org/wpilib/snippets/controlstutorialexamples/Robot.java
@@ -2,9 +2,9 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package org.wpilib.examples.controlstutorialexamples;
+package org.wpilib.snippets.controlstutorialexamples;
 
-import org.wpilib.examples.controlstutorialexamples.mechanisms.*;
+import org.wpilib.snippets.controlstutorialexamples.mechanisms.*;
 import org.wpilib.framework.RobotBase;
 import org.wpilib.framework.TimedRobot;
