+++ wpilibjExamples/src/main/java/org/wpilib/snippets/controlstutorialexamples/mechanisms/FlywheelPIDF.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package org.wpilib.examples.controlstutorialexamples.mechanisms;
+package org.wpilib.snippets.controlstutorialexamples.mechanisms;
 
 import org.wpilib.hardware.motor.PWMSparkMax;
 import org.wpilib.hardware.rotation.Encoder;
@@ -31,12 +31,12 @@ public class FlywheelPIDF implements AutoCloseable {
   private int kMotorPort = 0;
   private int kEncoderAChannel = 0;
   private int kEncoderBChannel = 1;
-  private double kEncoderDistPerPulse = 2.0 * Math.PI / 2048.0;
+  private double kFlywheelRadiansPerEncoderPulse = 2.0 * Math.PI / 2048.0;
   private DCMotor m_flywheelMotor;
   private Encoder m_encoder;
   private PWMSparkMax m_motor;
 
-  // Control Hardware: WPILib controllers for velocity control
+  // Controls Helpers: WPILib built-in classes for velocity control
   private PIDController m_controller;
   private SimpleMotorFeedforward m_feedforward;
 
@@ -53,7 +53,7 @@ public class FlywheelPIDF implements AutoCloseable {
   public FlywheelPIDF() {
     // Set up quadrature encoder for velocity measurement
     m_encoder = new Encoder(kEncoderAChannel, kEncoderBChannel);
-    m_encoder.setDistancePerPulse(kEncoderDistPerPulse);
+    m_encoder.setDistancePerPulse(kFlywheelRadiansPerEncoderPulse);
     
     // Set up SPARK PWM motor controller
     m_motor = new PWMSparkMax(kMotorPort);
