+++ wpilibjExamples/src/main/java/org/wpilib/snippets/controlstutorialexamples/mechanisms/TurretPositionPIDF.java
@@ -2,55 +2,44 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package org.wpilib.examples.controlstutorialexamples.mechanisms;
+package org.wpilib.snippets.controlstutorialexamples.mechanisms;
 
 import org.wpilib.hardware.motor.PWMSparkMax;
 import org.wpilib.hardware.rotation.Encoder;
-import org.wpilib.math.controller.ElevatorFeedforward;
-import org.wpilib.math.controller.ProfiledPIDController;
+import org.wpilib.math.controller.PIDController;
 import org.wpilib.math.system.DCMotor;
-import org.wpilib.math.trajectory.TrapezoidProfile;
 import org.wpilib.simulation.BatterySim;
-import org.wpilib.simulation.ElevatorSim;
 import org.wpilib.simulation.EncoderSim;
 import org.wpilib.simulation.PWMMotorControllerSim;
 import org.wpilib.simulation.RoboRioSim;
+import org.wpilib.simulation.SingleJointedArmSim;
 import org.wpilib.smartdashboard.SmartDashboard;
 import org.wpilib.system.RobotController;
 
-public class ElevatorPIDF implements AutoCloseable {
+public class TurretPositionPIDF implements AutoCloseable {
   // Physical mechanism constants
-  private double kGearing = 10.0;
-  private double kDrumRadius = 0.0508; // 2 inches in meters
-  private double kCarriageMass = 4.0; // kg
-  private double kMinHeight = 0.0; // m
-  private double kMaxHeight = 1.25; // m
+  private double kGearing = 300.0;
+  private double kMomentOfInertia = 0.005; // kg * m^2
 
   // Tuned Controller Constants - Tune these like in the tutorial!
   private double kKp = 5.0;
   private double kKi = 0.0;
   private double kKd = 0.0;
-  private double kProfileMaxVelocity = 2.45; // m/s
-  private double kProfileMaxAcceleration = 2.45; // m/s^2
-  private double kElevatorKs = 0.0; // static friction
-  private double kElevatorKg = 0.762; // gravity
-  private double kElevatorKv = 0.762; // velocity
 
   // Electronics Hardware: CIM motor controlled via SPARK PWM motor controller
-  private int kMotorPort = 3;
-  private int kEncoderAChannel = 6;
-  private int kEncoderBChannel = 7;
-  private double kEncoderDistPerPulse = 2.0 * Math.PI * 0.0508 / 4096.0; // 2" drum
-  private DCMotor m_elevatorMotor;
+  private int kMotorPort = 1;
+  private int kEncoderAChannel = 2;
+  private int kEncoderBChannel = 3;
+  private double kTurretRadiansPerEncoderPulse = 2.0 * Math.PI / 2048.0;
+  private DCMotor m_turretMotor;
   private Encoder m_encoder;
   private PWMSparkMax m_motor;
 
-  // Control Hardware: WPILib controllers for position control with motion profiling
-  private ProfiledPIDController m_controller;
-  private ElevatorFeedforward m_feedforward;
+  // Controls Helpers: WPILib built-in classes for position control
+  private PIDController m_controller;
 
   // Simulation Support
-  private ElevatorSim m_elevatorSim;
+  private SingleJointedArmSim m_turretSim;
   private EncoderSim m_encoderSim;
   private PWMMotorControllerSim m_motorSim;
 
@@ -59,46 +48,38 @@ public class ElevatorPIDF implements AutoCloseable {
   private double m_voltage = 0.0;
   private double m_actualPosition = 0.0;
 
-  public ElevatorPIDF() {
+  public TurretPositionPIDF() {
     // Set up quadrature encoder for position measurement
     m_encoder = new Encoder(kEncoderAChannel, kEncoderBChannel);
-    m_encoder.setDistancePerPulse(kEncoderDistPerPulse);
+    m_encoder.setDistancePerPulse(kTurretRadiansPerEncoderPulse);
     
     // Set up SPARK PWM motor controller
     m_motor = new PWMSparkMax(kMotorPort);
     
-    // Set up profiled PID controller for position control with motion profiling
-    m_controller = new ProfiledPIDController(
-        kKp,
-        kKi,
-        kKd,
-        new TrapezoidProfile.Constraints(kProfileMaxVelocity, kProfileMaxAcceleration));
+    // Set up WPILib's built-in PID controller for position control
+    m_controller = new PIDController(kKp, kKi, kKd);
     
-    // Set up elevator feedforward controller for gravity and velocity compensation
-    m_feedforward = new ElevatorFeedforward(kElevatorKs, kElevatorKg, kElevatorKv, 0.0);
   }
 
   // Initialize simulation components 
   /**
    * Initializes the simulation components for this mechanism.
    * This method should be called when running in simulation mode to set up
-   * the physics simulation models for the elevator mechanism.
+   * the physics simulation models for the turret mechanism.
    */
   public void initializeSimulation() {
     // Set up CIM motor model for simulation
-    m_elevatorMotor = DCMotor.getCIM(1);
+    m_turretMotor = DCMotor.getCIM(1);
     
-    // Set up simulation model for the elevator mechanism
-    m_elevatorSim = new ElevatorSim(
-        m_elevatorMotor,
+    // Set up simulation model for the turret mechanism
+    m_turretSim = new SingleJointedArmSim(
+        m_turretMotor,
         kGearing,
-        kCarriageMass,
-        kDrumRadius,
-        kMinHeight,
-        kMaxHeight,
+        kMomentOfInertia,
+        0.1, // sample length
+        -Math.PI, // full rotation range
+        Math.PI,
         true,
-        0.0,
-        0.01,
         0.0);
     
     // Set up simulation model for the encoder
@@ -109,10 +90,10 @@ public class ElevatorPIDF implements AutoCloseable {
   }
 
   /**
-   * Updates the elevator control loop.
+   * Updates the turret control loop.
    * This method should be called periodically (typically every 20ms) to:
    * 1. Read the current position from the encoder
-   * 2. Calculate the profiled PID + feedforward control output
+   * 2. Calculate the PID + feedforward control output
    * 3. Send the calculated voltage to the motor
    */
   public void update() {
@@ -120,19 +101,16 @@ public class ElevatorPIDF implements AutoCloseable {
     m_actualPosition = m_encoder.getDistance();
 
     // Step 2: Calculate
-    m_controller.setGoal(m_desiredPosition);
-    double pidOutput = m_controller.calculate(m_actualPosition);
-    double feedforwardOutput = m_feedforward.calculate(m_controller.getSetpoint().velocity);
-    m_voltage = pidOutput + feedforwardOutput;
+    m_voltage = m_controller.calculate(m_actualPosition, m_desiredPosition);
 
     // Step 3: Send Outputs
     m_motor.setVoltage(m_voltage);
   }
 
   /**
-   * Sets the desired position setpoint for the elevator.
+   * Sets the desired position setpoint for the turret.
    *
-   * @param setpoint The desired position in meters
+   * @param setpoint The desired position in radians
    */
   public void setSetpoint(double setpoint) {
     m_desiredPosition = setpoint;
@@ -144,12 +122,12 @@ public class ElevatorPIDF implements AutoCloseable {
    * to update the physics simulation and synchronize simulated sensors.
    */
   public void updateSimulation() {
-    if (m_elevatorSim != null) {
-      m_elevatorSim.setInput(m_motorSim.getThrottle() * RobotController.getBatteryVoltage());
-      m_elevatorSim.update(0.020);
-      m_encoderSim.setDistance(m_elevatorSim.getPosition());
+    if (m_turretSim != null) {
+      m_turretSim.setInput(m_motorSim.getThrottle() * RobotController.getBatteryVoltage());
+      m_turretSim.update(0.020);
+      m_encoderSim.setDistance(m_turretSim.getAngle());
       RoboRioSim.setVInVoltage(
-          BatterySim.calculateDefaultBatteryLoadedVoltage(m_elevatorSim.getCurrentDraw()));
+          BatterySim.calculateDefaultBatteryLoadedVoltage(m_turretSim.getCurrentDraw()));
     }
   }
 
@@ -159,9 +137,9 @@ public class ElevatorPIDF implements AutoCloseable {
    * information for debugging and monitoring.
    */
   public void updateTelemetry() {
-    SmartDashboard.putNumber("ElevatorPIDF/MotorVoltage", m_voltage);
-    SmartDashboard.putNumber("ElevatorPIDF/ActualPosition", m_actualPosition);
-    SmartDashboard.putNumber("ElevatorPIDF/DesiredPosition", m_desiredPosition);
+    SmartDashboard.putNumber("TurretPositionPIDF/MotorVoltage", m_voltage);
+    SmartDashboard.putNumber("TurretPositionPIDF/ActualPosition", m_actualPosition);
+    SmartDashboard.putNumber("TurretPositionPIDF/DesiredPosition", m_desiredPosition);
   }
 
   /**
