+++ wpilibjExamples/src/main/java/org/wpilib/snippets/controlstutorialexamples/mechanisms/VerticalArmPositionPIDF.java
@@ -2,11 +2,12 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package org.wpilib.examples.controlstutorialexamples.mechanisms;
+package org.wpilib.snippets.controlstutorialexamples.mechanisms;
 
 import org.wpilib.hardware.motor.PWMSparkMax;
 import org.wpilib.hardware.rotation.Encoder;
 import org.wpilib.math.controller.PIDController;
+import org.wpilib.math.controller.SimpleMotorFeedforward;
 import org.wpilib.math.system.DCMotor;
 import org.wpilib.simulation.BatterySim;
 import org.wpilib.simulation.EncoderSim;
@@ -16,30 +17,32 @@ import org.wpilib.simulation.SingleJointedArmSim;
 import org.wpilib.smartdashboard.SmartDashboard;
 import org.wpilib.system.RobotController;
 
-public class TurretPositionPIDF implements AutoCloseable {
+public class VerticalArmPositionPIDF implements AutoCloseable {
   // Physical mechanism constants
-  private double kGearing = 300.0;
-  private double kMomentOfInertia = 0.005; // kg * m^2
+  private double kGearing = 200.0;
+  private double kMomentOfInertia = 0.01; // kg * m^2
 
   // Tuned Controller Constants - Tune these like in the tutorial!
   private double kKp = 5.0;
   private double kKi = 0.0;
   private double kKd = 0.0;
+  private double kKf = 0.0; // position feedforward gain
 
   // Electronics Hardware: CIM motor controlled via SPARK PWM motor controller
-  private int kMotorPort = 1;
-  private int kEncoderAChannel = 2;
-  private int kEncoderBChannel = 3;
-  private double kEncoderDistPerPulse = 2.0 * Math.PI / 2048.0;
-  private DCMotor m_turretMotor;
+  private int kMotorPort = 2;
+  private int kEncoderAChannel = 4;
+  private int kEncoderBChannel = 5;
+  private double kVerticalArmRadiansPerEncoderPulse = 2.0 * Math.PI / 2048.0;
+  private DCMotor m_armMotor;
   private Encoder m_encoder;
   private PWMSparkMax m_motor;
 
-  // Control Hardware: WPILib controllers for position control
+  // Controls Helpers: WPILib built-in classes for position control
   private PIDController m_controller;
+  private SimpleMotorFeedforward m_feedforward;
 
   // Simulation Support
-  private SingleJointedArmSim m_turretSim;
+  private SingleJointedArmSim m_armSim;
   private EncoderSim m_encoderSim;
   private PWMMotorControllerSim m_motorSim;
 
@@ -48,10 +51,10 @@ public class TurretPositionPIDF implements AutoCloseable {
   private double m_voltage = 0.0;
   private double m_actualPosition = 0.0;
 
-  public TurretPositionPIDF() {
+  public VerticalArmPositionPIDF() {
     // Set up quadrature encoder for position measurement
     m_encoder = new Encoder(kEncoderAChannel, kEncoderBChannel);
-    m_encoder.setDistancePerPulse(kEncoderDistPerPulse);
+    m_encoder.setDistancePerPulse(kVerticalArmRadiansPerEncoderPulse);
     
     // Set up SPARK PWM motor controller
     m_motor = new PWMSparkMax(kMotorPort);
@@ -59,21 +62,23 @@ public class TurretPositionPIDF implements AutoCloseable {
     // Set up WPILib's built-in PID controller for position control
     m_controller = new PIDController(kKp, kKi, kKd);
     
+    // Set up WPILib's built-in feedforward controller (no velocity feedforward for position control)
+    m_feedforward = new SimpleMotorFeedforward(0.0, kKf, 0.0);
   }
 
   // Initialize simulation components 
   /**
    * Initializes the simulation components for this mechanism.
    * This method should be called when running in simulation mode to set up
-   * the physics simulation models for the turret mechanism.
+   * the physics simulation models for the vertical arm mechanism.
    */
   public void initializeSimulation() {
     // Set up CIM motor model for simulation
-    m_turretMotor = DCMotor.getCIM(1);
+    m_armMotor = DCMotor.getCIM(1);
     
-    // Set up simulation model for the turret mechanism
-    m_turretSim = new SingleJointedArmSim(
-        m_turretMotor,
+    // Set up simulation model for the vertical arm mechanism
+    m_armSim = new SingleJointedArmSim(
+        m_armMotor,
         kGearing,
         kMomentOfInertia,
         0.1, // sample length
@@ -90,7 +95,7 @@ public class TurretPositionPIDF implements AutoCloseable {
   }
 
   /**
-   * Updates the turret control loop.
+   * Updates the vertical arm control loop.
    * This method should be called periodically (typically every 20ms) to:
    * 1. Read the current position from the encoder
    * 2. Calculate the PID + feedforward control output
@@ -101,14 +106,16 @@ public class TurretPositionPIDF implements AutoCloseable {
     m_actualPosition = m_encoder.getDistance();
 
     // Step 2: Calculate
-    m_voltage = m_controller.calculate(m_actualPosition, m_desiredPosition);
+    double pidOutput = m_controller.calculate(m_actualPosition, m_desiredPosition);
+    double feedforwardOutput = m_feedforward.calculate(0.0); // No velocity feedforward for position control
+    m_voltage = pidOutput + feedforwardOutput;
 
     // Step 3: Send Outputs
     m_motor.setVoltage(m_voltage);
   }
 
   /**
-   * Sets the desired position setpoint for the turret.
+   * Sets the desired position setpoint for the vertical arm.
    *
    * @param setpoint The desired position in radians
    */
@@ -122,12 +129,12 @@ public class TurretPositionPIDF implements AutoCloseable {
    * to update the physics simulation and synchronize simulated sensors.
    */
   public void updateSimulation() {
-    if (m_turretSim != null) {
-      m_turretSim.setInput(m_motorSim.getThrottle() * RobotController.getBatteryVoltage());
-      m_turretSim.update(0.020);
-      m_encoderSim.setDistance(m_turretSim.getAngle());
+    if (m_armSim != null) {
+      m_armSim.setInput(m_motorSim.getThrottle() * RobotController.getBatteryVoltage());
+      m_armSim.update(0.020);
+      m_encoderSim.setDistance(m_armSim.getAngle());
       RoboRioSim.setVInVoltage(
-          BatterySim.calculateDefaultBatteryLoadedVoltage(m_turretSim.getCurrentDraw()));
+          BatterySim.calculateDefaultBatteryLoadedVoltage(m_armSim.getCurrentDraw()));
     }
   }
 
@@ -137,9 +144,9 @@ public class TurretPositionPIDF implements AutoCloseable {
    * information for debugging and monitoring.
    */
   public void updateTelemetry() {
-    SmartDashboard.putNumber("TurretPositionPIDF/MotorVoltage", m_voltage);
-    SmartDashboard.putNumber("TurretPositionPIDF/ActualPosition", m_actualPosition);
-    SmartDashboard.putNumber("TurretPositionPIDF/DesiredPosition", m_desiredPosition);
+    SmartDashboard.putNumber("VerticalArmPositionPIDF/MotorVoltage", m_voltage);
+    SmartDashboard.putNumber("VerticalArmPositionPIDF/ActualPosition", m_actualPosition);
+    SmartDashboard.putNumber("VerticalArmPositionPIDF/DesiredPosition", m_desiredPosition);
   }
 
   /**
