+++ wpilibjExamples/src/main/java/org/wpilib/snippets/snippets.json
@@ -332,5 +332,23 @@
     "foldername": "solenoid",
     "gradlebase": "java",
     "robotclass": "Robot"
+  },
+  {
+    "name": "Controls Tutorial Examples",
+    "description": "Demonstrates various control mechanisms: Flywheel (Bang-Bang and PIDF), Turret Position PIDF, Vertical Arm Position PIDF, and Elevator PIDF with motion profiling.",
+    "tags": [
+      "Controls",
+      "PID",
+      "Feedforward",
+      "Motion Profiling",
+      "Flywheel",
+      "Turret",
+      "Arm",
+      "Elevator",
+      "Simulation"
+    ],
+    "foldername": "controlstutorialexamples",
+    "gradlebase": "java",
+    "robotclass": "Robot"
   }
 ]
